// types.ts
export interface Service {
  title: string;
  icon: string;
  description: string;
}

export interface FAQItem {
  question: string;
  answer: string;
}

export interface VideoReview {
  id: string;
  title: string;
  thumbnail: string;
}

export interface ValueProp {
  title: string;
  description: string;
  icon: string;
}